Get most of the routes working
